testdata/htmltodo/box-sizing-valid.htm

<!DOCTYPE html>
<html><head>
<meta charset="utf-8">
<title>CSS UI Level 3: parsing box-sizing with valid values</title>
<link href="mailto:ericwilligers@chromium.org" rel="author" title="Eric Willigers">
<link href="https://drafts.csswg.org/css-ui-3/#box-sizing" rel="help">
<meta content="box-sizing supports the full grammar 'content-box | border-box'." name="assert">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("box-sizing", "content-box");
test_valid_value("box-sizing", "border-box");
</script>


</body></html>
